add constructor check
