+++ pom.xml
@@ -27,11 +27,11 @@
             <version>3.1</version>
         </dependency>
 
-        <dependency>
-            <groupId>commons-collections</groupId>
-            <artifactId>commons-collections</artifactId>
-            <version>3.2.1</version>
-        </dependency>
+        <!--<dependency>-->
+            <!--<groupId>commons-collections</groupId>-->
+            <!--<artifactId>commons-collections</artifactId>-->
+            <!--<version>3.2.1</version>-->
+        <!--</dependency>-->
 
         <dependency>
             <groupId>junit</groupId>
