+++ src/main/java/org/klose/scheme/constant/SConstant.java
@@ -38,5 +38,7 @@ public class SConstant {
     public static final String TRUE_SYMBOL = "true";
     public static final String FALSE_SYMBOL = "false";
 
-
+    private SConstant() {
+        throw new UnsupportedOperationException("illegal constructor for Sconstant");
+    }
 }
