+++ src/main/java/org/klose/scheme/utils/SParser.java
@@ -8,34 +8,34 @@ import java.util.*;
 
 /**
  * parse input into several SExpression
- *
+ * <p>
  * (define add (lambda (x y) (+ x y)) -> SExpression S0
- *
- *                            +-----------+
- *                            |           |
- *                            |     S0    |
- *                            +-----+-----+
- *                                  |
- *      +--------------------------------------------------------+
- *      |                           |                            |
+ * <p>
+ * +-----------+
+ * |           |
+ * |     S0    |
+ * +-----+-----+
+ * |
+ * +--------------------------------------------------------+
+ * |                           |                            |
  * +----v-----+                +----v-----+                +-----v----+
  * |          |                |          |                |          |
  * | "define" |                |   "add"  |                |    S1    |
  * +----------+                +----------+                +----------+
- *                                                               |
- *                               +------------------------------------
- *                               |                  |                |
- *                          +----v-----+      +-----v----+     +-----v-----+
- *                          |          |      |          |     |           |
- *                          | "lambda" |      |   S11    |     |    S12    |
- *                          +----------+      +----------+     +-----------+
- *                                                  |                |
- *                      +----------------+-----------                +----------------+----------------+
- *                      |                |                           |                |                |
- *                 +----v-----+    +-----v------+              +-----v-----+    +-----v-----+    +-----v-----+
- *                 |          |    |            |              |           |    |           |    |           |
- *                 |   "x"    |    |     "y"    |              |    "+"    |    |    "x"    |    |    "y"    |
- *                 +----------+    +------------+              +-----------+    +-----------+    +-----------+
+ * |
+ * +------------------------------------
+ * |                  |                |
+ * +----v-----+      +-----v----+     +-----v-----+
+ * |          |      |          |     |           |
+ * | "lambda" |      |   S11    |     |    S12    |
+ * +----------+      +----------+     +-----------+
+ * |                |
+ * +----------------+-----------                +----------------+----------------+
+ * |                |                           |                |                |
+ * +----v-----+    +-----v------+              +-----v-----+    +-----v-----+    +-----v-----+
+ * |          |    |            |              |           |    |           |    |           |
+ * |   "x"    |    |     "y"    |              |    "+"    |    |    "x"    |    |    "y"    |
+ * +----------+    +------------+              +-----------+    +-----------+    +-----------+
  */
 
 public class SParser {
@@ -101,7 +101,7 @@ public class SParser {
     public static void main(String[] args) {
 
         Scanner input = new Scanner(System.in);
-        List<String> lines = new ArrayList<String>();
+        List<String> lines = new ArrayList<>();
         String lineNew;
 
         while (input.hasNextLine()) {
@@ -117,5 +117,9 @@ public class SParser {
         for (String string : lines) {
             System.out.println(string);
         }
+    }
+
+    private SParser() {
+        throw new UnsupportedOperationException("illegal constructor for SParser");
     }
 }
